task #1: add typescript to unit-test
