+++ package.json
@@ -11,6 +11,7 @@
     "@types/node": "^16.18.108",
     "@types/react": "^18.3.8",
     "@types/react-dom": "^18.3.0",
+    "@types/redux-mock-store": "^1.5.0",
     "@ya.praktikum/react-developer-burger-ui-components": "^1.14.2",
     "react": "^18.3.1",
     "react-dnd": "^16.0.1",
