+++ src/services/reducers/ingredients.test.ts
@@ -1,43 +1,44 @@
 import configureMockStore from "redux-mock-store";
-import reducer, {
-  setIngredients,
-} from "./ingredients";
+import reducer, { setIngredients } from "./ingredients";
+import { TIngredient } from "utils/constants/types";
 
 const mockStore = configureMockStore();
-const payloadIngredients = [
+const payloadIngredients: TIngredient[] = [
   {
-    _id: 'test',
-    name: 'test',
-    type: 'bun',
+    _id: "test",
+    name: "test",
+    type: "bun",
     proteins: 1000,
     fat: 1000,
     carbohydrates: 1000,
     calories: 3000,
     price: 1111,
-    image: 'http://test',
-    image_mobile: 'http://test',
-    image_large: 'http://test',
-    __v: '1111111',
+    image: "http://test",
+    image_mobile: "http://test",
+    image_large: "http://test",
+    __v: 1111111,
   },
   {
-    _id: 'test1',
-    name: 'test1',
-    type: 'sauce',
+    _id: "test1",
+    name: "test1",
+    type: "sauce",
     proteins: 10001,
     fat: 10001,
     carbohydrates: 10001,
     calories: 30001,
     price: 111112,
-    image: 'http://test1',
-    image_mobile: 'http://test1',
-    image_large: 'http://test1',
-    __v: '11111112',
+    image: "http://test1",
+    image_mobile: "http://test1",
+    image_large: "http://test1",
+    __v: 11111112,
   },
 ];
 
 describe("ingredients reducer", () => {
   it("should return the initial state", () => {
-    expect(reducer(undefined, {})).toEqual([]);
+    const unknownAction = { type: "unknown" };
+
+    expect(reducer(undefined, unknownAction)).toEqual([]);
   });
 
   it("should handle setIngredients action", () => {
@@ -51,7 +52,6 @@ describe("ingredients reducer", () => {
     const store = mockStore([]);
     store.dispatch(setIngredients(payloadIngredients));
 
-    expect(store.getActions()[0])
-      .toEqual(setIngredients(payloadIngredients));
+    expect(store.getActions()[0]).toEqual(setIngredients(payloadIngredients));
   });
 });
